C_Rules entfernt und einen Großteil der Methoden in die Default Klasse übernommen.
